backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
sqlalchemy==2.0.48
aiosqlite==0.22.1
python-multipart==0.0.22
PyMuPDF==1.27.2.2
openpyxl==3.1.5
openai==2.30.0
python-dotenv==1.2.2
pydantic==2.12.5
Pillow==11.2.1
pytesseract==0.3.13
json5==0.9.25
pytest==8.4.1
httpx==0.28.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
beautifulsoup4==4.12.3
python-docx==1.1.2
boto3==1.35.90
celery==5.4.0
redis==5.2.1
qdrant-client==1.12.1
requests==2.32.3
requests-oauthlib==2.0.0
